+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- JEI support for compounders (metal alloyer and recrystallizer)
+
 ## [2.6.1] - 2021-02-18
 ### Added
 - Tyrian Steel: a new metal alloy and gear material